backend/requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
sqlmodel==0.0.22
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
pydantic==2.10.0
pydantic-settings==2.7.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.20
cryptography==43.0.3
httpx==0.28.1
